requirements.txt
asgiref==3.4.1
attrs==23.1.0
black==22.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cleo==2.0.1
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.6
django-cors-headers==3.8.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
drf-yasg==1.20.0
dulwich==0.21.3
filelock==3.12.0
html5lib==1.1
idna==3.3
importlib-metadata==6.6.0
inflection==0.5.1
installer==0.7.0
isort==5.10.1
itypes==1.2.0
jaraco.classes==3.2.3
Jinja2==3.0.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.41
Pillow==9.0.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
psycopg2==2.9.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
shellingham==1.5.0.post1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.4.25
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.21.1
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
